refactor : Card에 필요한 prop만 전달하도록 변경
