+++ src/components/feature/CartSection/CartList/index.tsx
@@ -2,6 +2,9 @@ import { CartProduct } from "../../../../type/cart";
 import CheckBox from "../../../common/CheckBox";
 import Card from "../CartProducts/Card";
 import styled from "@emotion/styled";
+import { deleteCartProduct } from "../../../../api/cart/deleteCartProduct";
+import { updateCartProduct } from "../../../../api/cart/updateCartProduct";
+import { useShowError } from "../../../../provider/errorProvider";
 
 type Props = {
   cartItems: CartProduct[] | undefined;
@@ -16,6 +19,8 @@ const CartList = ({
   setSelectedCartIds,
   refetch,
 }: Props) => {
+  const showError = useShowError();
+
   const isChecked = (id: number) => {
     return selectedCartIds?.some((item: number) => item === id);
   };
@@ -40,9 +45,24 @@ const CartList = ({
     setSelectedCartIds(selectedCartIds?.filter((cartId) => cartId !== id));
   };
 
-  const handleDelete = (id: number) => {
+  const handleDelete = async (id: number) => {
     if (!selectedCartIds?.find((item: number) => item === id)) return;
     setSelectedCartIds(selectedCartIds?.filter((cartId) => cartId !== id));
+    try {
+      await deleteCartProduct(id);
+      refetch();
+    } catch (e) {
+      showError?.("데이터를 삭제하는 중 문제가 발생했습니다.");
+    }
+  };
+
+  const handleUpdate = async (id: number, updatedQuantity: number) => {
+    try {
+      await updateCartProduct(id, updatedQuantity);
+      refetch();
+    } catch (e) {
+      showError?.("상품을 추가/삭제하는 중 문제가 발생했습니다.");
+    }
   };
 
   return (
@@ -58,10 +78,10 @@ const CartList = ({
           <Card
             key={cartItem.id}
             cartItem={cartItem}
-            onRefetch={refetch}
             isChecked={isChecked(cartItem.id)}
             onToggle={() => handleToggle(cartItem.id)}
-            onDeleteSelected={() => handleDelete(cartItem.id)}
+            onDelete={(id: number) => handleDelete(id)}
+            onUpdate={(id: number, count: number) => handleUpdate(id, count)}
           />
         ))}
       </CardList>
