+++ src/components/feature/CartSection/CartProducts/Card/index.tsx
@@ -4,48 +4,17 @@ import Button from "../../../../common/Button";
 import CheckBox from "../../../../common/CheckBox";
 import Line from "../../../../common/Line";
 import { CartProduct } from "../../../../../type/cart";
-import { deleteCartProduct } from "../../../../../api/cart/deleteCartProduct";
 import { formatPrice } from "../../../../../utils/formatPrice";
-import { useShowError } from "../../../../../provider/errorProvider";
-import { updateCartProduct } from "../../../../../api/cart/updateCartProduct";
-
 type Props = {
   cartItem: CartProduct;
   isChecked: boolean;
-  onRefetch: () => void;
   onToggle: () => void;
-  onDeleteSelected: () => void;
+  onDelete: (id: number) => void;
+  onUpdate: (id: number, count: number) => void;
 };
 
-const Card = ({
-  cartItem,
-  onRefetch,
-  isChecked,
-  onToggle,
-  onDeleteSelected,
-}: Props) => {
+const Card = ({ cartItem, isChecked, onToggle, onDelete, onUpdate }: Props) => {
   const { imageUrl, name, price } = cartItem.product;
-  const showError = useShowError();
-
-  const handleDelete = async (id: number) => {
-    try {
-      await deleteCartProduct(id);
-      onDeleteSelected();
-      onRefetch();
-    } catch (e) {
-      showError?.("데이터를 삭제하는 중 문제가 발생했습니다.");
-    }
-  };
-
-  const handleUpdate = async (id: number, updatedQuantity: number) => {
-    try {
-      await updateCartProduct(id, updatedQuantity);
-      onRefetch();
-      console.log(cartItem.quantity);
-    } catch (e) {
-      showError?.("상품을 추가/삭제하는 중 문제가 발생했습니다.");
-    }
-  };
 
   return (
     <S.CardContainer>
@@ -55,7 +24,7 @@ const Card = ({
           onChange={onToggle}
           testId={`check-box${cartItem.id}`}
         />
-        <Button onClick={() => handleDelete(cartItem.id)} title="삭제" />
+        <Button onClick={() => onDelete(cartItem.id)} title="삭제" />
       </S.ButtonSection>
 
       <S.CardInfoSection>
@@ -71,10 +40,8 @@ const Card = ({
           </S.ProductDescription>
           <CartCount
             count={cartItem.quantity}
-            onPlusCount={() => handleUpdate(cartItem.id, cartItem.quantity + 1)}
-            onMinusCount={() =>
-              handleUpdate(cartItem.id, cartItem.quantity - 1)
-            }
+            onPlusCount={() => onUpdate(cartItem.id, cartItem.quantity + 1)}
+            onMinusCount={() => onUpdate(cartItem.id, cartItem.quantity - 1)}
             testId={`count${cartItem.id}`}
           />
         </S.ProductInfoSection>
